feat: Add work experience section
This commit introduces a new section to the site for the work experience. This includes the creation a new component called "JobCard", new pages called "jobs" in the partials folder (including the spanish version), a a new "job" folder inside the content folder with new markdown files. The Navigation component is modified to reflect this new section and the "Aside" component adds updated links to the resume files.