refactor: Add new props to JobCard and ProjectCard components
This commit introduces a new prop called 'techLegend' for the JobCard and ProjectCard components as well for the 'job' and 'project' markdown files for both languages (es & en). The corresponding 'jobs' and 'projects' pages reflect this change. It also adds some changes to the overall style of the page (new background and border colors).