feat: Add CertificateCard component
This commit introduces a new section in the "Education" page. Now the course and certificates are separated. The Education page gets content from the new "certificates" content folder and imports the CertificateCard component to showcase all the certificates (currently one) in the page. The sames changes apply for the spanish version of the Education page.

Additionally, the Aside and Modal components receive some minor changes regarding some of the props names.