refactor: Add jobs description and tasks as Markdown content
This commit refactors the way the description and tasks for a job content file are injected into the JobCard component. Before, these description and tasks values were part of the Markdown file frontmatter. Now these description and tasks values are part of the content of the Markdown file. In order to render the Markdown content into HTML, the 'marked' package is installed.

Now the JobCard component renders Markdown content as HTML content. In order to apply styles to the Markdown content, it is necessary to override Tailwind's Preflight setting by introducing a global.css file with custom styling and import this CSS file into the BaseLayout component.

The jobs.astro partials reflect this change by passing "job.body" as props to the JobCard component instead of "job.description" and "job.tasks".