feat: Remove Contact pages | Refactor Social component
This commit removes the Contact pages from the project since the email address is now part of the Aside component. Due to this, the Form component is removed as well.

A function is added to the Social component frontmatter in order to parse the corresponding props and generate the component based on these props. This approach allows the Social component to generate a component for different social accounts based on the "platform" prop.